Added Requirements ability to UnitMetaData
Per key requirements can now be defined in the plugin's translator allowing values to be filtered/removed if they don't meet the proper validation.

### Example:

In this example, `'requires'` is applied to all `Category(environment.wind.direction)` since properties are inherited.  Per the specification this requirement restricts the logging of wind direction if the wind speed is not reported in the same observation or the wind speed is less than 0.

```python
{
  environment.wind.direction: {
    'type': 'direction',
    'requires': {
      environment.wind.speed.speed: {
        '>': 0
      }
   }
	environment.wind.direction.direction: {
    'title': 'Direction',
    'sourceUnit': 'º',
    'sourceKey': 'wind_direction'
  },
}
```