New Class: LevityDatagram
New class LevityDatagram, and it's subclass Subdatagram, are used for ingesting/interpreting raw data.  Raw data is mapped via dataMaps and keyMaps provided by the plugin's translator the correct map is either provided during initialization or inferred based on values whose keys are specified as metaData or sourceData in the plugin's translator.

Signed-off-by: Bryan Neal Garrison <git@noblecloud.io>